+++ data/i18n/hydrogen_es.ts
@@ -1464,6 +1464,16 @@ Please set your system&apos;s locale to UTF-8!</source>
         <extracomment>Text displayed on a Keep button of a dialog. The character after the &apos;&amp;&apos; symbol can be used as a hotkey and the &apos;&amp;&apos; symbol itself will not be displayed.</extracomment>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>Duplicate instrument</source>
+        <extracomment>Representing an instrument duplication in the undo history and in the popup menu of the sidebar of the pattern editor.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Pattern</source>
+        <extracomment>Indicates a menu section which affects patterns.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>ComponentView</name>
@@ -5855,10 +5865,6 @@ Are you sure?</source>
         <translation>Patrón guardado.</translation>
     </message>
     <message>
-        <source>Fill/Clear...</source>
-        <translation>Rellenar/Limpiar...</translation>
-    </message>
-    <message>
         <source>Could not save pattern to temporary directory.</source>
         <translation>No se pudo guardar el patrón en un directorio temporal.</translation>
     </message>
@@ -5876,6 +5882,14 @@ Are you sure?</source>
     </message>
     <message>
         <source>Replace</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Fill/Clear cells</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Select cells</source>
         <translation type="unfinished"></translation>
     </message>
 </context>