+++ src/gui/src/CommonStrings.h
@@ -249,6 +249,8 @@ class CommonStrings : public H2Core::Object<CommonStrings> {
 		const QString& getActionIrreversible() const { return m_sActionIrreversible; }
 
 		const QString& getActionAddInstrument() const { return m_sActionAddInstrument; }
+		const QString& getActionDuplicateInstrument() const {
+			return m_sActionDuplicateInstrument; }
 		const QString& getActionDeleteInstrument() const { return m_sActionDeleteInstrument; }
 		const QString& getActionDropInstrument() const { return m_sActionDropInstrument; }
 		const QString& getActionRenameInstrument() const { return m_sActionRenameInstrument; }
@@ -356,6 +358,7 @@ class CommonStrings : public H2Core::Object<CommonStrings> {
 		const QString& getErrorUniqueTypes() const {return m_sErrorUniqueTypes; }
 
 		const QString& getSettings() const { return m_sSettings; }
+		const QString& getPattern() const { return m_sPattern; }
 private:
 	QString m_sSmallSoloButton;
 	QString m_sSmallMuteButton;
@@ -551,6 +554,7 @@ private:
 		QString m_sActionIrreversible;
 
 		QString m_sActionAddInstrument;
+		QString m_sActionDuplicateInstrument;
 		QString m_sActionDeleteInstrument;
 		QString m_sActionDropInstrument;
 		QString m_sActionRenameInstrument;
@@ -627,5 +631,6 @@ private:
 	QString m_sEncodingError;
 
 		QString m_sSettings;
+		QString m_sPattern;
 };
 #endif