+++ src/gui/src/MainForm.h
@@ -190,8 +190,10 @@ public slots:
 
 		static void action_drumkit_addInstrument(
 			std::shared_ptr<H2Core::Instrument> pInstrument = nullptr );
+		static void action_drumkit_duplicateInstrument( int nInstrumentIndex );
 		static void action_drumkit_deleteInstrument( int nInstrumentIndex );
-		static void action_drumkit_renameInstrument( int nInstrumentIndex );
+		static void action_drumkit_renameInstrument( int nInstrumentIndex,
+													const QString& sNewName );
 
 		void action_window_showMixer();
 		void action_window_showPlaylistEditor();