+++ src/gui/src/SongEditor/SongEditor.h
@@ -83,6 +83,7 @@ class SongEditor : public Editor::Base<std::shared_ptr<GridCell>>
 		int getGridHeight() { return m_nGridHeight; }
 
 		void clearThePatternSequenceVector( const QString& sFileName );
+		void selectAllCellsInRow( int nRow );
 
 		int yScrollTarget( QScrollArea *pScrollArea, int *pnPatternInView );